Show product reviews in front store product-view

+++ src/main/java/com/github/queebskeleton/hardwarecommerce/entity/Rating.java
@@ -24,6 +24,6 @@ public class Rating {
 	
 	private String review;
 	
-	private byte value;
+	private int value;
 	
 }
